+++ Blaster-Master/GridManager.cpp
@@ -27,17 +27,17 @@ GridManager::GridManager()
 
 int GridManager::round(int num, int other)
 {
-	return (num % other == 0) ? num / other : ceil(num / other) + 1;
+	return (num % other == 0) ? num / other : (int) ceil(num / other) + 1;
 }
 
 void GridManager::Clear()
 {
-	for (int i = 0; i < grids.size(); i++)
+	for (size_t i = 0; i < grids.size(); i++)
 	{
 		grids[i].clear();
 	}
 	grids.clear();
-	for (int i = 0; i < objects.size(); i++)
+	for (size_t i = 0; i < objects.size(); i++)
 	{
 		delete objects[i];
 	}
@@ -78,20 +78,20 @@ void GridManager::_AddToGrid(LPGAMEOBJECT object)
 	float left, top, right, bottom;
 	object->GetBoundingBox(left, top, right, bottom);
 
-	float ceil_width = CGame::GetInstance()->GetScreenWidth() / 2.0;
-	float ceil_height = CGame::GetInstance()->GetScreenHeight() / 2.0;
+	float ceil_width = CGame::GetInstance()->GetScreenWidth() / 2.0f;
+	float ceil_height = CGame::GetInstance()->GetScreenHeight() / 2.0f;
 
-	int fromRow = bottom / ceil_height;
-	int toRow = top / ceil_height;
+	size_t fromRow = (size_t) (bottom / ceil_height);
+	size_t toRow = (size_t) (top / ceil_height);
 
-	int fromColumn = left / ceil_width;
-	int toColumn = right / ceil_width;
+	size_t fromColumn = (size_t) (left / ceil_width);
+	size_t toColumn = (size_t) (right / ceil_width);
 
 	if (fromRow < 0 || fromColumn < 0 || toRow > grids.size() - 1 || toColumn > grids[0].size() - 1) return;
 
-	for (int row = fromRow; row <= toRow; ++row)
+	for (size_t row = fromRow; row <= toRow; ++row)
 	{
-		for (int column = fromColumn; column <= toColumn; ++column)
+		for (size_t column = fromColumn; column <= toColumn; ++column)
 		{
 			grids.at(row).at(column)->AddObject(object);
 		}
@@ -102,12 +102,10 @@ void GridManager::_AddToGrid(LPGAMEOBJECT object)
 
 vector<LPGAMEOBJECT> GridManager::GetObjectsToUpdate()
 {
-
-
-	for (int row = 0; row < grids.size(); ++row)
+	for (size_t row = 0; row < grids.size(); ++row)
 	{
 		vector<Grid*> rowGrid = grids[row];
-		for (int column = 0; column < rowGrid.size(); ++column)
+		for (size_t column = 0; column < rowGrid.size(); ++column)
 		{
 			Grid* grid = rowGrid[column];
 			grid->ClearObjects();
@@ -123,20 +121,20 @@ vector<LPGAMEOBJECT> GridManager::GetObjectsToUpdate()
 		return obj->shouldRemove == true;
 	}), collisions.end());*/
 
-	for (int i = 0; i < objects.size(); ++i)
+	for (size_t i = 0; i < objects.size(); ++i)
 	{
 		_AddToGrid(objects[i]);
 	}
 
 	vector<LPGAMEOBJECT> result;
-	float ceil_width = CGame::GetInstance()->GetScreenWidth() / 2.0;
-	float ceil_height = CGame::GetInstance()->GetScreenHeight() / 2.0;
+	float ceil_width = CGame::GetInstance()->GetScreenWidth() / 2.0f;
+	float ceil_height = CGame::GetInstance()->GetScreenHeight() / 2.0f;
 
 	float x, y;
 	Camera::GetInstance()->GetPosition(x, y);
 
-	int fromColumn = x / ceil_width;
-	int toRow = y / ceil_height;
+	int fromColumn = (int)  (x / ceil_width);
+	int toRow = (int) (y / ceil_height);
 
 	int toColumn = fromColumn + 2 >= grids.at(0).size() - 1 ? grids.at(0).size() - 1 : fromColumn + 2;
 
@@ -146,7 +144,7 @@ vector<LPGAMEOBJECT> GridManager::GetObjectsToUpdate()
 		{
 			Grid* currentGrid = grids.at(row).at(column);
 			vector<LPGAMEOBJECT> objects = currentGrid->GetObjects();
-			for (int i = 0; i < objects.size(); ++i)
+			for (size_t i = 0; i < objects.size(); ++i)
 			{
 				LPGAMEOBJECT obj = objects[i];
 				if (dynamic_cast<Enemy*>(obj))
@@ -167,7 +165,7 @@ vector<LPGAMEOBJECT> GridManager::GetObjectsToUpdate()
 					}
 				}
 				bool isExist = false;
-				for (int index = 0; index < result.size(); ++index)
+				for (size_t index = 0; index < result.size(); ++index)
 				{
 					if (obj == result[index])
 					{
